requirements.txt
git+https://github.com/Pycord-Development/pycord
discord-rich-presence
python-dotenv